+++ destinations/faros-destination/src/converters/github/collaborators.ts
@@ -6,6 +6,7 @@ import {
   DestinationRecord,
   StreamName,
 } from '../converter';
+import {GithubCommon} from './common';
 
 export class GithubCollaborators implements Converter {
   readonly streamName = new StreamName('github', 'collaborators');
@@ -18,37 +19,6 @@ export class GithubCollaborators implements Converter {
     const source = this.streamName.source;
     const user = record.record.data;
 
-    const type = ((): {category: string; detail: string} => {
-      if (!user.type) {
-        return {category: 'Custom', detail: 'unknown'};
-      }
-      const userTypeLower = user.type.toLowerCase();
-      switch (userTypeLower) {
-        case 'enterpriseuseraccount':
-        case 'user':
-          return {category: 'User', detail: userTypeLower};
-        case 'bot':
-          return {category: 'Bot', detail: userTypeLower};
-        case 'organization':
-          return {category: 'Organization', detail: userTypeLower};
-        case 'mannequin':
-          return {category: 'Mannequin', detail: userTypeLower};
-        default:
-          return {category: 'Custom', detail: userTypeLower};
-      }
-    })();
-
-    return [
-      {
-        model: 'vcs_User',
-        record: {
-          uid: user.login,
-          name: user.name ?? null,
-          htmlUrl: user.html_url ?? null,
-          type,
-          source,
-        },
-      },
-    ];
+    return [GithubCommon.vcs_User(user, source)];
   }
 }
